Revert "Change script to be able to run against any network"
This reverts commit 2ec95722edab5830079861129bc81cb0998f84d1.

Encountering some problems with creating fixtures for tests.
Stick to hard coding fork url in the script
